Refactor renderer modules
